agent/requirements.txt
mysql-connector-python>=9.0.0
psutil>=6.0.0

